+++ src/Serializer.fs
@@ -16,11 +16,13 @@ module Serialize =
                 NullValueHandling = NullValueHandling.Include
             )
 
-        let serialize obj =
-            JsonConvert.SerializeObject (obj, options())
+        let serializeWith (options: JsonSerializerSettings) obj =
+            JsonConvert.SerializeObject (obj, options)
 
-        let serializePretty obj =
-            let options = options()
+        let serialize obj =
+            obj |> serializeWith (options())
+
+        let private serializePrettyWith (options: JsonSerializerSettings) obj =
             options.Formatting <- Formatting.Indented
 
             use stringWriter = new StringWriter()
@@ -32,8 +34,76 @@ module Serialize =
 
             stringWriter.ToString()
 
+        let serializePretty obj =
+            obj |> serializePrettyWith (options())
+
+        let private optionsIgnoringNulls () =
+            let options = options ()
+            options.NullValueHandling <- NullValueHandling.Ignore
+            options
+
+        let serializeIgnoringNulls obj =
+            obj |> serializeWith (optionsIgnoringNulls())
+
+        let serializeIgnoringNullsPretty obj =
+            obj |> serializePrettyWith (optionsIgnoringNulls())
+
     let toJsonPretty: obj -> string = Json.serializePretty
     let toJson: obj -> string = Json.serialize
+
+    let toJsonIgnoringNullsPretty: obj -> string = Json.serializeIgnoringNullsPretty
+    let toJsonIgnoringNulls: obj -> string = Json.serializeIgnoringNulls
+
+    [<RequireQualifiedAccess>]
+    module JsonValue =
+        open FSharp.Data
+        open System.Collections.Generic
+
+        /// Dictionary is used to maintain the order of added items - not sorting the keys like Map does.
+        let private dictionaryOfArray array =
+            array
+            |> Array.fold
+                (fun (acc: Dictionary<string, _>) (key, value) ->
+                    acc.Add(key, value)
+                    acc
+                )
+                (new Dictionary<string, _>())
+
+        let rec toSerializableJson: JsonValue -> obj = function
+            | JsonValue.String string -> string :> obj
+            | JsonValue.Number number -> number |> int :> obj
+            | JsonValue.Float float -> float :> obj
+            | JsonValue.Boolean bool -> bool :> obj
+            | JsonValue.Record properties ->
+                properties
+                |> Array.map (fun (key, value) -> key, value |> toSerializableJson)
+                |> dictionaryOfArray
+                :> obj
+            | JsonValue.Array values ->
+                values
+                |> Array.map toSerializableJson
+                :> obj
+            | JsonValue.Null -> null
+
+        let rec toSerializableJsonIgnoringNullsInRecord: JsonValue -> obj = function
+            | JsonValue.String string -> string :> obj
+            | JsonValue.Number number -> number |> int :> obj
+            | JsonValue.Float float -> float :> obj
+            | JsonValue.Boolean bool -> bool :> obj
+            | JsonValue.Record properties ->
+                // Json serializer ignores nulls in objects (records), but not in maps, dictionaries, etc. So we need to ignore them manually.
+                properties
+                |> Array.choose (function
+                    | _, JsonValue.Null -> None
+                    | key, value -> Some (key, value |> toSerializableJsonIgnoringNullsInRecord)
+                )
+                |> dictionaryOfArray
+                :> obj
+            | JsonValue.Array values ->
+                values
+                |> Array.map toSerializableJsonIgnoringNullsInRecord
+                :> obj
+            | JsonValue.Null -> null
 
     open System